% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{sva}
\alias{sva}
\title{A function for estimating surrogate variables by estimating empirical
control probes}
\usage{
sva(dat, mod, mod0 = NULL, n.sv=NULL, controls=NULL, 
    method =  c("irw", "two-step", "supervised"), vfilter = NULL,
    B = 5, numSVmethod = "be")
}
\arguments{
\item{dat}{The transformed data matrix with the variables in rows and samples
in columns}

\item{mod}{The model matrix being used to fit the data}

\item{mod0}{The null model being compared when fitting the data}

\item{n.sv}{The number of surogate variables to estimate}

\item{controls}{A vector of probabilities (between 0 and 1, inclusive) that
each gene is a control. A value of 1 means the gene is certainly a control
and a value of 0 means the gene is certainly not a control.}

\item{method}{For empirical estimation of control probes use "irw". If
control probes are known use "supervised"}

\item{vfilter}{You may choose to filter to the vfilter most variable rows
before performing the analysis. vfilter must be NULL if method is 
"supervised"}

\item{B}{The number of iterations of the irwsva algorithm to perform}

\item{numSVmethod}{If n.sv is NULL, sva will attempt to estimate the number
of needed surrogate variables. This should not be adapted by the user
unless they are an expert.}
}
\value{
sv The estimated surrogate variables, one in each column

pprob.gam: A vector of the posterior probabilities each gene is
    affected by heterogeneity

pprob.b A vector of the posterior probabilities each gene is
    affected by mod

n.sv The number of significant surrogate variables
}
\description{
This function is the implementation of the iteratively re-weighted least
squares approach for estimating surrogate variables. As a by product, this
function produces estimates of the probability of being an empirical control.
See the function \code{\link{empirical.controls}} for a direct estimate of
the empirical controls.
}
\examples{
library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:5000,]

pheno = pData(dat)
edata = exprs(dat)
mod = model.matrix(~as.factor(cancer), data=pheno)
mod0 = model.matrix(~1,data=pheno)

n.sv = num.sv(edata,mod,method="leek")
svobj = sva(edata,mod,mod0,n.sv=n.sv)

}
